reused mock load_given_json_list_data_from_tempdir_to_bq_mock
